pyproject.toml
[project]
name = "yt-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.195.0",
    "google-auth-oauthlib>=1.3.1",
    "pathlib>=1.0.1",
    "rich>=15.0.0",
    "typer>=0.25.1",
]
[project.scripts]
yt-cli="yt_cli.cli:app"